refactor(base-gen): extract StatusChangeMixin (Tier-1b Task 2)
Move the upgrade/downgrade status-change bucket out of the monolithic
BaseGenerationService ABC into a composed StatusChangeMixin, the cleanest
boundary of the Tier-1b decomposition (zero per-run self fields, zero
billing).

New package reflexio/server/services/base_generation/ holds:
- _status_change.py: StatusChangeMixin[TRequest] with the 8 methods moved
  verbatim (run_upgrade, run_downgrade, and the 6 overridable hooks
  _has_items_with_status, _delete_items_by_status, _update_items_status,
  _get_affected_user_ids_for_upgrade, _get_affected_user_ids_for_downgrade,
  _create_status_change_response). The StatusChangeOperation enum moves here
  too and is re-exported from base_generation_service for back-compat.
- __init__.py: re-exports StatusChangeMixin + StatusChangeOperation.

BaseGenerationService now composes StatusChangeMixin ahead of ABC. Behavior
preserving: module path and class path unchanged; __abstractmethods__ stays
the exact 8-name set; all ~30 subclass-contract names still resolve on the
3 subclasses via MRO; run_upgrade/run_downgrade remain public for the
lib/_profiles.py and lib/_user_playbook.py callers.
