+++ reflexio/server/services/base_generation/__init__.py
@@ -1,0 +1,18 @@
+"""Sub-mixins composed into ``BaseGenerationService`` (Tier-1b decomposition).
+
+Each mixin holds a cohesive bucket of the ``BaseGenerationService`` behaviour and
+is composed into the base via MRO in
+``reflexio.server.services.base_generation_service``. The residual base keeps the
+ABC skeleton (all abstract methods), ``__init__`` (per-run field inits), the ``run``
+FIFO drain, ``_run_generation`` billing emit-ordering, and the lock/queue helpers.
+"""
+
+from reflexio.server.services.base_generation._status_change import (
+    StatusChangeMixin,
+    StatusChangeOperation,
+)
+
+__all__ = [
+    "StatusChangeMixin",
+    "StatusChangeOperation",
+]
